// api/models/getListeningStats200Items.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Audiobookshelf API
 * Audiobookshelf API with autogenerated OpenAPI doc
 * OpenAPI spec version: 0.1.0
 */
import type { BookMetadata } from './bookMetadata';

/**
 * A list of library items the user has listened to, keyed by their item IDs.
 */
export type GetListeningStats200Items = {
  [key: string]: {
    /** The ID of the library item. */
    id?: string;
    /** The time (in seconds) the user listened to this library item. */
    timeListening?: number;
    mediaMetadata?: BookMetadata;
  };
};