// api/models/mediaProgress.ts
/**
 * Generated by orval v7.9.0 🍺
 * Do not edit manually.
 * Audiobookshelf API
 * Audiobookshelf API with autogenerated OpenAPI doc
 * OpenAPI spec version: 0.1.0
 */
import type { CreatedAt } from './createdAt';
import type { CurrentTime } from './currentTime';
import type { DurationSec } from './durationSec';
import type { FinishedAt } from './finishedAt';
import type { HideFromContinueListening } from './hideFromContinueListening';
import type { IsFinished } from './isFinished';
import type { LibraryItemId } from './libraryItemId';
import type { MediaProgressId } from './mediaProgressId';
import type { PodcastId } from './podcastId';
import type { Progress } from './progress';
import type { UpdatedAt } from './updatedAt';

export interface MediaProgress {
  id?: MediaProgressId;
  libraryItemId?: LibraryItemId;
  episodeId?: PodcastId;
  duration?: DurationSec;
  progress?: Progress;
  currentTime?: CurrentTime;
  isFinished?: IsFinished;
  hideFromContinueListening?: HideFromContinueListening;
  lastUpdate?: UpdatedAt;
  startedAt?: CreatedAt;
  finishedAt?: FinishedAt;
}